Updated user list to use DataGrid, updated selection of role - it integrates fully with both database and updates state on the frontend. Ability to remove the last sysadmin is still functional.
